refactor(ai-agent): move capability modules into capability/
