+++ crates/aionui-ai-agent/src/capability/backend_output_sink.rs
