+++ crates/aionui-ai-agent/src/capability/backend_protocol_sink.rs
