+++ crates/aionui-ai-agent/src/capability/cli_process.rs
