+++ crates/aionui-ai-agent/src/capability/skill_manager.rs
