+++ crates/aionui-ai-agent/src/capability/team_guide_prompt.rs
