+++ crates/aionui-ai-agent/src/lib.rs
@@ -8,11 +8,8 @@ pub mod agent_registry;
 pub mod agent_routes;
 pub mod agent_task;
 pub mod aionrs_agent;
-pub mod backend_output_sink;
-pub mod backend_protocol_sink;
-pub mod cli_process;
+pub mod capability;
 pub mod factory;
-pub mod first_message_injector;
 pub mod idle_scanner;
 pub mod manager;
 pub mod nanobot_agent;
@@ -20,10 +17,8 @@ pub mod openclaw;
 pub mod protocol;
 pub mod routes;
 pub mod shared_kernel;
-pub mod skill_manager;
 pub mod stream_event;
 pub mod task_manager;
-mod team_guide_prompt;
 pub mod types;
 
 pub use acp_agent::AcpAgentManager;
@@ -39,9 +34,14 @@ pub use aionui_api_types::{
     AcpBuildExtra, AcpModelInfo, AcpSessionConfigOption, AionrsBuildExtra, OpenClawBuildExtra, OpenClawGatewayConfig,
     RemoteBuildExtra, SlashCommandItem,
 };
-pub use backend_output_sink::BackendOutputSink;
-pub use backend_protocol_sink::BackendProtocolSink;
-pub use cli_process::CliAgentProcess;
+pub use capability::backend_output_sink::BackendOutputSink;
+pub use capability::backend_protocol_sink::BackendProtocolSink;
+pub use capability::cli_process::CliAgentProcess;
+pub use capability::skill_manager::{
+    AcpSkillManager, SkillDefinition, SkillIndex, build_skills_index_text, build_system_instructions,
+    build_system_instructions_with_skills_index, detect_skill_load_request, prepare_first_message,
+    prepare_first_message_with_skills_index,
+};
 pub use factory::{AgentFactoryDeps, build_agent_factory};
 pub use idle_scanner::start_idle_scanner;
 pub use manager::acp::AcpSessionSyncService;
@@ -51,11 +51,6 @@ pub use manager::remote::{
 pub use nanobot_agent::NanobotAgentManager;
 pub use openclaw::OpenClawAgentManager;
 pub use routes::{SessionRouterState, session_routes};
-pub use skill_manager::{
-    AcpSkillManager, SkillDefinition, SkillIndex, build_skills_index_text, build_system_instructions,
-    build_system_instructions_with_skills_index, detect_skill_load_request, prepare_first_message,
-    prepare_first_message_with_skills_index,
-};
 pub use stream_event::AgentStreamEvent;
 pub use task_manager::{AgentFactory, IWorkerTaskManager, WorkerTaskManagerImpl};
 pub use types::{AgentStreamChunk, AionrsCompatOverrides, AionrsResolvedConfig, BuildTaskOptions, SendMessageData};
