+++ crates/aionui-ai-agent/src/nanobot_agent.rs
@@ -9,7 +9,7 @@ use tracing::{debug, error, info, warn};
 
 use aionui_common::CommandSpec;
 
-use crate::cli_process::CliAgentProcess;
+use crate::capability::cli_process::CliAgentProcess;
 use crate::stream_event::AgentStreamEvent;
 use crate::types::SendMessageData;
 use std::path::PathBuf;
