+++ crates/aionui-ai-agent/src/openclaw/agent.rs
@@ -9,7 +9,7 @@ use tokio::sync::{Mutex, RwLock, broadcast};
 use tracing::{debug, error, info, warn};
 
 use crate::agent_manager::approval_key;
-use crate::cli_process::CliAgentProcess;
+use crate::capability::cli_process::CliAgentProcess;
 use crate::stream_event::AgentStreamEvent;
 use crate::types::SendMessageData;
 use aionui_api_types::{OpenClawBuildExtra, OpenClawGatewayConfig};
